backend/requirements.txt
requests
pandas
msal
python-dotenv
openpyxl
